🚚 api: [PRESCRIPTION] move file to prescription/campaign context
