+++ api/src/prescription/campaign/infrastructure/serializers/csv/campaign-profiles-collection-export.js
@@ -5,8 +5,8 @@ import {
   CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING,
   CONCURRENCY_HEAVY_OPERATIONS,
 } from '../../../../../../lib/infrastructure/constants.js';
-import { CampaignProfilesCollectionResultLine } from '../../../../../../lib/infrastructure/exports/campaigns/campaign-profiles-collection-result-line.js';
 import * as csvSerializer from '../../../../../../lib/infrastructure/serializers/csv/csv-serializer.js';
+import { CampaignProfilesCollectionResultLine } from '../../exports/campaigns/campaign-profiles-collection-result-line.js';
 class CampaignProfilesCollectionExport {
   constructor(outputStream, organization, campaign, competences, translate) {
     this.stream = outputStream;
